put all of the database classes and methods into one class. After review I will add javadoc comments.
